requirements.txt
# AI Trading System - Pinned Dependencies
# Install with: pip install -r requirements.txt

# Core
pandas==3.0.0
numpy==2.4.1
requests==2.32.5
python-dotenv==1.2.1
python-dateutil==2.9.0.post0

# Web Server (for frontend API)
flask==3.0.0
flask-cors==4.0.0

# Technical Analysis
TA-Lib==0.6.8

# Data APIs
ccxt==4.5.38

# Dashboard & Visualization
plotly==6.5.2
dash==4.0.0

# News & Sentiment
newsapi-python==0.2.7
textblob==0.19.0

# Machine Learning
scikit-learn==1.8.0

# Async (real-time)
aiohttp==3.13.3
websockets==16.0

# Testing
pytest==9.0.2
pytest-cov==6.0.0
flake8==7.0.0

# Code Quality
pylint==4.0.4
black==25.1.0

# Utilities
tqdm==4.67.3

# FastAPI (for API server)
fastapi==0.115.0
uvicorn==0.34.0
python-multipart==0.0.20
pydantic-settings==2.7.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
